+++ CHANGELOG.md
@@ -4,6 +4,12 @@
 FEATURES:
 - Added logic to create content from template files via the new files submodule
 - Content generated dynamically, using the highly opinionated configurations.
+- The following content is managed by the bootstrap_
+- - PR template
+- - Support template
+- - terraform-docs workflow and configuration
+- - Semver and Release workflow
+- - Examples skeleton with configured auth.tf and documentation examples section - TBD
 
 BUG FIXES:
 - Fixed wrong key generation for labels, secrets and vars